auth_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
asyncpg
alembic
python-dotenv
